Reduce extension logging
The previous logging was quite verbose, and messages were logged
every time a website was opened.
Additionally it could have been a privacy risk because the domain and
URL were logged, even for Incognito windows. And the extension log
could be retrieved retroactively if the browser was still open.

With the new approach verbose logging can be enabled on demand,
if needed.

A few messages are still logged by default:
- messages at extension start (before the `_KNOWN_SITES_DEBUG` flag can be set)
- messages for interacting with the extension site (without including domain or URL)
- messages for warnings and errors, which should normally not occur