Allow requests with tabId -1 for Chrome
Relates to #69
However, this does not fully fix this because Chrome seems to send
multiple implicit requests and only some have tabId -1.
Other requests, including those for `https://www.google.com/search` and
the URL for the site the user has searched (even without trying to open
the site), have the real tabId, causing them to be blocked, which is
quite confusing.